move some classes around
