+++ src/main/java/com/cleanroommc/modularui/widget/SlotWidget.java
@@ -6,7 +6,7 @@ import com.cleanroommc.modularui.api.math.Pos2d;
 import com.cleanroommc.modularui.api.math.Size;
 import com.cleanroommc.modularui.drawable.IDrawable;
 import com.cleanroommc.modularui.drawable.UITexture;
-import com.cleanroommc.modularui.slot.BaseSlot;
+import com.cleanroommc.modularui.integration.vanilla.slot.BaseSlot;
 import net.minecraft.inventory.Slot;
 import org.apache.commons.lang3.ArrayUtils;
